+++ spirit-log/examples/background.rs
@@ -20,12 +20,12 @@ mod everything {
     #[derive(Clone, Debug, StructOpt)]
     struct Opts {
         #[structopt(flatten)]
-        log: LogOpts,
+        logging: LogOpts,
     }
 
     impl Opts {
-        fn log(&self) -> LogOpts {
-            self.log.clone()
+        fn logging(&self) -> LogOpts {
+            self.logging.clone()
         }
     }
 
@@ -37,14 +37,14 @@ mod everything {
 
     #[derive(Clone, Debug, Default, Deserialize)]
     struct Cfg {
-        #[serde(flatten)]
-        log: LogCfg,
+        #[serde(default, skip_serializing_if = "LogCfg::is_empty")]
+        logging: LogCfg,
         ui: Ui,
     }
 
     impl Cfg {
-        fn log(&self) -> LogCfg {
-            self.log.clone()
+        fn logging(&self) -> LogCfg {
+            self.logging.clone()
         }
     }
 
@@ -71,8 +71,8 @@ mod everything {
             .with(
                 Pipeline::new("logging")
                     .extract(|opts: &Opts, cfg: &Cfg| LogBoth {
-                        cfg: cfg.log(),
-                        opts: opts.log(),
+                        cfg: cfg.logging(),
+                        opts: opts.logging(),
                     })
                     // Transforms the logger to the background one.
                     .transform(Background::new(100, OverflowMode::Block)),